At this version has been added:
The incoming real time data are more organized and programmed to return open, closed, high, low and volume.

The real time data are combined with the historical data and printed as a data frame in console.

There is the need to correct the process of adding the real time data to the data frame as there is some delay in comparison with the data that are returned by the requests.